requirements.txt
# fastapi==0.115.6
# paddleocr==2.9.1
# paddlepaddle==2.6.2
# paho-mqtt==2.1.0
# pandas==2.2.3
# pillow==11.0.0
# redis==5.2.1
# opencv-python-headless==4.10.0.84  # Thay vì opencv-contrib-python nếu không cần các modules bổ sung
# cloudinary==1.42.0
# ultralytics==8.3.50
# uvicorn==0.34.0
# numpy==1.26.4
# python-multipart
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.42.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
# opencv-contrib-python==4.10.0.84
# opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.0.0
pip==23.0.1
protobuf==3.20.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
redis==5.2.1
requests==2.32.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
setuptools==65.5.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.50
ultralytics-thop==2.0.13
urllib3==2.2.3
uvicorn==0.34.0